# Loop benchmark for more "stable" results