--- a/src/test/java/com/pngencoder/PngEncoderBenchmarkAssorted.java
+++ b/src/test/java/com/pngencoder/PngEncoderBenchmarkAssorted.java
@@ -29,6 +29,8 @@ public class PngEncoderBenchmarkAssorted {
         }
     }
 
+    private final static boolean DISALBE_IMAGE_IO = false;
+
     @Disabled("run manually")
     @Test
     public void runBenchmarkCustom() throws IOException {
@@ -40,25 +42,32 @@ public class PngEncoderBenchmarkAssorted {
 
         final File outImageIO = File.createTempFile("out-imageio", ".png");
         //final File outPngEncoder = File.createTempFile("out-pngencoder", ".png");
-        final File outPngEncoder = new File("/Users/olof/Desktop/out.png");
+        final File outPngEncoder = new File("target/test/assorted_out.png");
+        System.out.println(outPngEncoder.getAbsolutePath());
+        outPngEncoder.getParentFile().mkdir();
 
-        ImageIO.write(original, "png", outImageIO);
-        Timing.message("ImageIO Warmup");
+        if (!DISALBE_IMAGE_IO) {
+            ImageIO.write(original, "png", outImageIO);
+            Timing.message("ImageIO Warmup");
 
-        ImageIO.write(original, "png", outImageIO);
-        Timing.message("ImageIO Result");
+            ImageIO.write(original, "png", outImageIO);
+            Timing.message("ImageIO Result");
+        }
 
         PngEncoder pngEncoder = new PngEncoder()
-                //.withMultiThreadedCompressionEnabled(false)
-                .withCompressionLevel(9)
+                .withMultiThreadedCompressionEnabled(true)
+                //.withPredictorEncoding(true)
+                .withCompressionLevel(4)
                 .withBufferedImage(original);
         System.out.println(outPngEncoder);
 
         pngEncoder.toFile(outPngEncoder);
         Timing.message("PngEncoder Warmup");
 
-        pngEncoder.toFile(outPngEncoder);
-        Timing.message("PngEncoder Result");
+        for (int i = 0; i < 10; i++) {
+            pngEncoder.toFile(outPngEncoder);
+            Timing.message("PngEncoder Result " + i);
+        }
 
         final long imageIOSize = outImageIO.length();
         final long pngEncoderSize = outPngEncoder.length();