@johantiden 2022-04-11
Nice work on the benchmarks and in splitting up PRs into separate concerns!